# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SiteCraft AI Website Generation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for current user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, generating, ready, deployed, failed]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project metadata
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get AI job status and progress
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Job status and progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIJob"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /prompts:
    get:
      operationId: listPrompts
      tags: [prompts]
      summary: List all prompt templates in the library
      responses:
        "200":
          description: List of prompt templates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptListResponse"

    post:
      operationId: createPrompt
      tags: [prompts]
      summary: Create a new prompt template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptRequest"
      responses:
        "201":
          description: Prompt created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
  /prompts/{id}:
    patch:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update a prompt template (creates a new version)
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromptRequest"
      responses:
        "200":
          description: Prompt updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        email:
          type: string
        profileImage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
    Project:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        businessDescription:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, generating, ready, deployed, failed]
        theme:
          type: string
          nullable: true
        previewUrl:
          type: string
          nullable: true
        liveUrl:
          type: string
          nullable: true
        generatedHtml:
          type: string
          nullable: true
        designTokens:
          nullable: true
        seoScore:
          type: number
          nullable: true
        accessibilityScore:
          type: number
          nullable: true
        performanceScore:
          type: number
          nullable: true
        visualScore:
          type: number
          nullable: true
        activeJobId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - status
        - createdAt
        - updatedAt
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - projects
        - total
        - page
        - limit
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        businessDescription:
          type: string
      required:
        - name
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    AIJob:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        type:
          type: string
          enum: [generate, chat-edit, regenerate-section]
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        progress:
          type: number
          description: 0-100
        currentStep:
          type: string
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/AIJobStep"
        error:
          type: string
          nullable: true
        result:
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - projectId
        - type
        - status
        - progress
        - steps
        - createdAt
        - updatedAt
    AIJobStep:
      type: object
      properties:
        id:
          type: string
        jobId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        order:
          type: integer
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        output:
          nullable: true
        error:
          type: string
          nullable: true
      required:
        - id
        - jobId
        - name
        - status
        - order
...
    Prompt:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        agentRole:
          type: string
        description:
          type: string
        systemPrompt:
          type: string
        userPromptTemplate:
          type: string
        model:
          type: string
          enum: [gemini-flash, gemini-pro]
        temperature:
          type: number
        version:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - agentRole
        - description
        - systemPrompt
        - userPromptTemplate
        - model
        - temperature
        - version
        - isActive
        - createdAt
        - updatedAt
    PromptListResponse:
      type: object
      properties:
        prompts:
          type: array
          items:
            $ref: "#/components/schemas/Prompt"
      required:
        - prompts
    CreatePromptRequest:
      type: object
      properties:
        name:
          type: string
        agentRole:
          type: string
        description:
          type: string
        systemPrompt:
          type: string
        userPromptTemplate:
          type: string
        model:
          type: string
          enum: [gemini-flash, gemini-pro]
        temperature:
          type: number
      required:
        - name
        - agentRole
        - systemPrompt
        - userPromptTemplate
        - model
    UpdatePromptRequest:
      type: object
      properties:
        name:
          type: string
        systemPrompt:
          type: string
        userPromptTemplate:
          type: string
        model:
          type: string
          enum: [gemini-flash, gemini-pro]
        temperature:
          type: number
        changelog:
          type: string
...